% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{ACF}
\alias{ACF}
\title{ACF Plot for Markov chain Monte Carlo}
\usage{
ACF(x,component = NULL, type = c("correlation", "covariance"),
             plot= TRUE, lag.max = NULL, avg.col = "blue", chain.col   = "red",
             na.action   = na.fail, auto.layout = TRUE, ask = dev.interactive())
}
\arguments{
\item{x}{: an `Smcmc' class object or a list of Markov chains or a Markov chain matrix}

\item{component}{: a vector of integers indicating which components' ACF plots are needed. By default all components are drawn.}

\item{type}{: the kind of ACF plot: "correlation" or "covariance"}

\item{plot}{: TRUE if plots are required. If FALSE, raw values are returned}
  
 \item{main}{: To add heading ro the plot}

\item{lag.max}{: Maximum lag for the ACF plot}

\item{chain.col}{: color for the ACF of the individual chains.}

\item{na.action}{:  function to be called to handle missing values. ‘na.pass’ can be used.}

\item{auto.layout}{: logical argument for an automatic layout of plots}

\item{ask}{: activating interactive plots}

\item{avg.col}{: color for the overall ACF of each component}
}
\value{
returns the autocorrelation function plots of the Markov chains. Uses the
        more accurate globally-centered ACFs.
}
\description{
Autocorrelation function plots for MCMC data (including multiple chains)
}
\examples{
# Producing Markov chain
chain <- matrix(0, ncol = 1, nrow = 1e3)
chain[1,] <- 0
err <- rnorm(1e3)
for(i in 2:1e3)
{
  chain[i,] <- .3*chain[i-1,] + err[i]
}
chain <- Smcmc(list(chain))
ACF(chain)

}
\references{
Agarwal, M., and Vats, D., “Globally-centered autocovariances in MCMC”, 
arxiv - 2009.01799,  2020.
}
