% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{acfplot}
\alias{acfplot}
\title{ACF Plot for Markov chain Monte Carlo}
\usage{
acfplot(x,which = NULL, type = c("correlation"),
               plot = TRUE, main = NA, xlab = "Lag", lag.max = NULL, avg.col = "blue", chain.col   = "red",
               na.action = na.fail, ...)
}
\arguments{
\item{x}{: an `Smcmc' class object or a list of Markov chains or a Markov chain matrix}

\item{which}{: a vector of integers indicating which components' ACF plots are needed. By default all components are drawn.}

\item{type}{: the kind of ACF plot: "correlation" or "covariance". By default it is "correlation"}

\item{plot}{: TRUE if plots are required. If FALSE, raw values are returned}

\item{main}{: main heading of plot}

\item{xlab}{: By default "Lag", pass another value, if you want to change.}

\item{lag.max}{: Maximum lag for the ACF plot}

\item{avg.col}{: color for the overall ACF of each component}

\item{chain.col}{: color for the ACF of the individual chains.}

\item{na.action}{:  function to be called to handle missing values. ‘na.pass’ can be used.}

\item{...}{: Other arguments}
}
\value{
returns the autocorrelation function plots of the Markov chains. Uses the
        more accurate globally-centered ACFs.
}
\description{
Autocorrelation function plots for MCMC data (including multiple chains)
}
\examples{
# Producing Markov chain
chain <- matrix(0, ncol = 1, nrow = 1e3)
chain[1,] <- 0
err <- rnorm(1e3)
for(i in 2:1e3)
{
  chain[i,] <- .3*chain[i-1,] + err[i]
}
chain <- Smcmc(list(chain))
acfplot(chain)

}
\references{
Agarwal, M., and Vats, D., “Globally-centered autocovariances in MCMC”, 
arxiv - 2009.01799,  2020.
}
