% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{addCI}
\alias{addCI}
\title{Add simultaneous confidence interval to existing plot.}
\usage{
addCI(x, CIs, component = 1, bord = NA, mean = TRUE, mean.color = 'plum4',
               quan.color = 'lightsteelblue3', opaq = 0.7, ...)
}
\arguments{
\item{x}{: a `Smcmc' class object}

\item{CIs}{: the output from the `getCI` function}

\item{component}{: numeric indicating which component to draw the confidence intervals for}

\item{bord}{: logical for whether a border is desired around the confidence intervals}

\item{mean}{: logical argument whether the mean is to be plotted}

\item{mean.color}{: color for the mean confidence interval}

\item{quan.color}{: color for the quantile confidence intervals}

\item{opaq}{: opacity of \code{mean.col} and \code{quan.col}. A value of 0 is transparent and 1 is completely opaque.}

\item{...}{: arguments passed on to the boundaries of the confidence intervals in `segments`}
}
\value{
adds segments for confidence intervals into an already existing plot environment
}
\description{
Adds simultaneous confidence intervals for quantiles and means to an existing plot.
}
\examples{
chain <- matrix(0, ncol = 1, nrow = 1e3)
chain[1,] <- 0
err <- rnorm(1e3)
for(i in 2:1e3)
{
  chain[i,] <- .3*chain[i-1,] + err[i]
}
chain <- Smcmc(list(chain))
plot(density(chain$stacked[,1]))
CIs <- getCI(chain)
addCI(chain, CIs, component = 1)
}
