% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QQ.R
\name{allocateP}
\alias{allocateP}
\title{Distribution of Negative Log P}
\usage{
allocateP(p, U = NULL, B = 1000)
}
\arguments{
\item{p}{Vector of p values.}

\item{U}{Upper limit of discretization grid.}

\item{B}{Bins.}
}
\description{
Count the number of p values exceeding each grid point
on the \eqn{-log_{10}} scale.
}
