% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/args.R
\name{as.fraction}
\alias{as.fraction}
\title{Convert a text as "2/3" to numeric fraction}
\usage{
as.fraction(x)
}
\arguments{
\item{x}{Text like "2/3}
}
\description{
Convert a text as "2/3" to numeric fraction
}
