% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{boxCI}
\alias{boxCI}
\title{Add simultaneous confidence interval to existing boxplot}
\usage{
boxCI(x, CI, component = c(1), dimn = 1, 
               quan.color = 'lightsteelblue3', horizontal = FALSE)
}
\arguments{
\item{x}{: a `Smcmc' class object}

\item{CI}{: the output from the `getCI` function with `iid = TRUE`}

\item{component}{: vector indicating which components to draw the confidence intervals for}

\item{dimn}{: numeric for how many components are being plotted}

\item{quan.color}{: color for the quantile confidence intervals}

\item{horizontal}{: logical for whether boxplots are horizontal}
}
\value{
adds segments for confidence intervals into an already existing plot environment
}
\description{
Adds simultaneous confidence intervals for quantiles to an existing boxplot.
}
\examples{
output <- matrix(rnorm(3*1e3), nrow = 1e3, ncol = 3)

}
