% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Characterize.R
\name{Characterize}
\alias{Characterize}
\title{Characterize Cumulants}
\usage{
Characterize(y, sig = 0.05, report = T)
}
\arguments{
\item{y}{Observations.}

\item{sig}{Significance level.}

\item{report}{Display result?}
}
\value{
Data.frame containing the estimated mean, variance, skewness, and kurtosis.
}
\description{
Estimates the cumulants of a distribution using a vector of observations.
}
\examples{
\dontrun{
# Normal sample
y = rnorm(n=1e3);
# Characterize
K = Characterize(y=y);
}
}
