% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Means.R
\name{comp.means}
\alias{comp.means}
\title{Compare Means}
\usage{
comp.means(y1, y0, indep = F, sig = 0.05)
}
\arguments{
\item{y1}{Realizations at time $t_{1}$.}

\item{y0}{Realizations at time $t_{0}$.}

\item{indep}{Are the measurements independent?}

\item{sig}{Significance level for CI.}
}
\description{
Given two measurements on a single group, or measurements on two different
groups, calculates the difference and ratio of means.
}
