% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Compare.R
\name{compCumulants}
\alias{compCumulants}
\title{Compare Cumulants}
\usage{
compCumulants(y1, y0, compare = "means", indep = F, sig = 0.05)
}
\arguments{
\item{y1}{Measurements at time $t_{1}$.}

\item{y0}{Measurements at time $t_{0}$.}

\item{compare}{Either means or vars.}

\item{indep}{Are the measurements independent?}

\item{sig}{Significance level for CI.}
}
\description{
Given two measurements on a single group, or measurements on two different 
groups, compares the group means or variances.
}
