% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Compare.R
\name{compEfficiency}
\alias{compEfficiency}
\title{Compare Efficiency}
\usage{
compEfficiency(p1, p0, df = 1, indep = F, sig = 0.05)
}
\arguments{
\item{p1}{P-values from test 1.}

\item{p0}{P-values from test 0.}

\item{df}{Degress of freedom, if comparing the efficiency of hypothesis
tests.}

\item{indep}{Are the measurements independent?}

\item{sig}{Significance level for CI.}
}
\description{
Given p-values from two tests of the same hypothesis on the same data,
or on two different data sets, compares the efficiency using the 
difference and ratio of non-centrality parameters.
}
