% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Smcmcclass.R
\name{convert2Smcmc}
\alias{convert2Smcmc}
\title{Covert to Smcmc Object}
\usage{
convert2Smcmc(x)
}
\arguments{
\item{x}{: a object belongs from any of "mcmc.list", "stanfit", "rstan", "array", "matrix" classes.}
}
\value{
return Smcmc object having same chain(s)
}
\description{
To covert different MCMC objects to Smcmc object
}
