% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Smcmcclass.R
\name{densityplot}
\alias{densityplot}
\title{density plot form Smcmc class}
\usage{
densityplot(x, Q = c(0.1, 0.9), alpha = 0.05, thresh = 0.001, iid = FALSE,
                            plot = TRUE, mean = TRUE, border = NA, mean.col = 'plum4', 
                            quan.col = 'lightsteelblue3',rug = TRUE, opaq = 0.7, 
                            auto.layout = TRUE, ask = dev.interactive(),...)
}
\arguments{
\item{x}{: a `Smcmc' class object}

\item{Q}{: vector of quantiles}

\item{alpha}{: confidence level of simultaneous confidence intervals}

\item{thresh}{: numeric typically less than .005 for the accuracy of the simulteaneous procedure}

\item{main}{: To add main heading}

\item{iid}{: logical argument for constructing density plot for iid samples. Defaults to \code{FALSE}}

\item{plot}{:  logical argument for is plots are to be returned}

\item{mean}{: logical argument whether the mean is to be plotted}

\item{which}{: A vector of components, if you want plots of specific components.}

\item{border}{: whether a border is required for the simultaneous confidence intervals}

\item{mean.col}{: color for the mean confidence interval}

\item{quan.col}{: color for the quantile confidence intervals}

\item{rug}{: logical indicating whether a rug plot is desired}

\item{opaq}{: opacity of \code{mean.col} and \code{quan.col}. A value of 0 is transparent and 1 is completely opaque.}

\item{auto.layout}{: logical argument for an automatic layout of plots}

\item{ask}{: activating inter active plots}

\item{fast}{: Argument passed to getCI}

\item{...}{: arguments passed on to the \code{density} plot in base R}
}
\value{
returns a plot of the univariate density estimates with simultaneous
		confidence intervals wherever asked. If \code{plot == FALSE} a list of
		estimates and simultaneous confidence intervals.
}
\description{
Density plots with simultaenous error bars around means and quantiles
 for MCMC data. The error bars account for the correlated nature of the process.
}
\examples{
# Producing Markov chain
chain <- matrix(0, ncol = 1, nrow = 1e3)
chain[1,] <- 0
err <- rnorm(1e3)
for(i in 2:1e3)
{
  chain[i,] <- .3*chain[i-1,] + err[i]
}
chain <- Smcmc(list(chain))
densityplot(chain)

}
\references{
Robertson, N., Flegal, J. M., Vats, D., and Jones, G. L., 
“Assessing and Visualizing Simultaneous Simulation Error”, 
Journal of Computational and Graphical Statistics,  2020.
}
