% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadGeno.R
\name{DrawGeno}
\alias{DrawGeno}
\title{Draw Genotypes from Plink}
\usage{
DrawGeno(m = 1, stem, IID = NULL)
}
\arguments{
\item{m}{Number of loci to import.}

\item{stem}{Plink stem BED, BIM, FAM files.}

\item{IID}{Individuals to select.}
}
\description{
Reads a random block of $m$ snps from a plink data set.
}
