% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/exampleData.R
\docType{data}
\name{exampleData}
\alias{exampleData}
\title{Simple example of including data}
\format{A data.frame with 10 rows and two columns:
\describe{
 \item{ID}{Unique identity variable}
 \item{Value}{Simulated value (g)}
 }}
\source{
Simulated data
}
\usage{
exampleData
}
\description{
This is a simple example of how we can include data in a package
and provide the corresponding documentation.
}
\keyword{datasets}

