% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadGeno.R
\name{GenoSeg}
\alias{GenoSeg}
\title{Read Genotype Block from Plink}
\usage{
GenoSeg(L = 1, U = NULL, stem, IID = NULL)
}
\arguments{
\item{L}{Beginning of segment.}

\item{U}{End of segment}

\item{stem}{Plink stem BED, BIM, FAM files.}

\item{IID}{Individuals to select.}
}
\description{
Reads a segment of genotypes from a plink file.
}
