% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/args.R
\name{get.param.closure}
\alias{get.param.closure}
\title{Get parameter by name from a grid}
\usage{
get.param.closure(i, grid)
}
\arguments{
\item{i}{Task ID}

\item{grid}{Parameter grid}
}
\description{
Get parameter by name from a grid
}
