% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{getCI}
\alias{getCI}
\title{Calculates simultaneous confidence intervals.}
\usage{
getCI(x, Q = c(0.1, 0.9), alpha = 0.05, thresh = 0.001, iid = FALSE,
            mean = TRUE)
}
\arguments{
\item{x}{: a `Smcmc' class object}

\item{Q}{: vector of quantiles}

\item{alpha}{: confidence levels of the simulatenous intervals}

\item{thresh}{: threshold for the optimization methodology that calculates the simultaneous CIs}

\item{iid}{: logical argument for constructing density plot for iid samples. Defaults to \code{FALSE}}

\item{mean}{: logical  indicating whether mean is to be plotted}

\item{fast}{: By defaults TRUE (To calculate CI efficiently), if set to FALSE, then use traditional method for CI calculations}
}
\value{
adds segments for confidence intervals into an already existing plot environment
}
\description{
Calculates simultaneous confidence intervals for means and
             quantiles as indicated for the desired MCMC output
}
\examples{
library(mvtnorm)
chain <- matrix(0, ncol = 1, nrow = 1e3)
chain[1,] <- 0
err <- rnorm(1e3)
for(i in 2:1e3)
{
  chain[i,] <- .3*chain[i-1,] + err[i]
}
chain <- Smcmc(list(chain))
plot(density(chain$stacked[,1]))
CIs <- getCI(chain)
addCI(chain, CIs, component = 1)

}
\references{
Robertson, N., Flegal, J. M., Vats, D., and Jones, G. L., 
“Assessing and Visualizing Simultaneous Simulation Error”, 
Journal of Computational and Graphical Statistics,  2020.
}
