% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kurtosis.R
\name{Kurtosis}
\alias{Kurtosis}
\title{Estimate Kurtosis}
\usage{
Kurtosis(y, k0 = 3, sig = 0.05, simple = F)
}
\arguments{
\item{y}{Observations.}

\item{k0}{Reference kurtosis.}

\item{sig}{Significance level.}

\item{simple}{If TRUE, returns estimated size and SE only.}
}
\value{
Data.frame containing the estimated kurtosis, its standard error, the
  lower and upper confidence bounds, and a p value assessing the null
  hypothesis that the kurtosis is \code{k0}.
}
\description{
Estimates the sample kurtosis using a vector of observations.
}
\examples{
\dontrun{
# Normal sample
y = rnorm(n=1e3);
# Test kurtosis = 3
Kurtosis(y=y,k0=3);
}
}
