% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NCP.R
\name{Ncp}
\alias{Ncp}
\title{Estimate Non-Centrality Parameter}
\usage{
Ncp(p, df = 1, sig = 0.05)
}
\arguments{
\item{p}{Vector of p-values.}

\item{df}{Degrees of freedom.}

\item{sig}{Significance level for CI.}
}
\description{
Estimates the non-centrality parameter of the chi-square distribution using a
vector of p-values.
}
