% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/args.R
\name{parse.args}
\alias{parse.args}
\title{Parse arguments from the command line, possibly in vector form}
\usage{
parse.args(arg, type_fun)
}
\arguments{
\item{arg}{Arguments, comma separated}

\item{type_fun}{Which type the argument should be}
}
\description{
Parse arguments from the command line, possibly in vector form
}
\examples{
parse.args("hello,world", as.character)
parse.args("1,2,3", as.integer)

}
