% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Siidclass.R
\name{plot.Siid}
\alias{plot.Siid}
\title{Plot Siid}
\usage{
\method{plot}{Siid}(x, which = NULL, Q = c(0.1, 0.9), alpha = 0.05, thresh = 0.001, 
                           rug = TRUE, plot = TRUE, mean = TRUE, border = NA,
                           mean.col = 'plum4', quan.col = 'lightsteelblue3', 
                           opaq = 0.7, auto.layout = TRUE, 
                           ask = dev.interactive(),main = NA,...)
}
\arguments{
\item{x}{: a `Siid' class object}

\item{which}{: if one want to plots specific components}

\item{Q}{: vector of quantiles}

\item{alpha}{: confidence level of simultaneous confidence intervals}

\item{thresh}{: numeric typically less than .005 for the accuracy of the simulteaneous procedure}

\item{rug}{: logical indicating whether a rug plot is desired}

\item{plot}{:  logical argument for is plots are to be returned}

\item{mean}{: logical argument whether the mean is to be plotted}

\item{border}{: whether a border is required for the simultaneous confidence intervals}

\item{mean.col}{: color for the mean confidence interval}

\item{quan.col}{: color for the quantile confidence intervals}

\item{opaq}{: opacity of \code{mean.col} and \code{quan.col}. A value of 0 is transparent and 1 is completely opaque.}

\item{auto.layout}{: logical argument for an automatic layout of plots}

\item{ask}{: activating interactive plots}

\item{main}{: for main heading}

\item{...}{: arguments passed on to the \code{density} plot in base R}
}
\value{
returns a plot of the univariate density estimates with simultaneous
     confidence intervals wherever asked. If \code{plot == FALSE} a list of
     estimates and simultaneous confidence intervals.
}
\description{
Density plots with simultaenous error bars around means and quantiles
 for iid data.
}
\examples{
# Generating iid data
chain <- matrix(rnorm(3*1e3), nrow = 1e3, ncol = 3)
siid.obj <- Siid(chain)
plot(siid.obj)

}
\references{
Robertson, N., Flegal, J. M., Vats, D., and Jones, G. L., 
“Assessing and Visualizing Simultaneous Simulation Error”, 
Journal of Computational and Graphical Statistics,  2020.
}
