% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QQ.R
\name{qqFrame}
\alias{qqFrame}
\title{Generate QQ Frame}
\usage{
qqFrame(breaks, counts)
}
\arguments{
\item{breaks}{Breakpoints.}

\item{counts}{Number of test statistics exceeding the breakpoint.}
}
\value{
A data.frame containing observed and expected uniform quantiles.
}
\description{
Generates points for the QQ plot. Assumes that \eqn{\beta_{0}=0}, and the
corresponding count is the total number of observations. Confidence intervals
are constructed using the beta distribution.
}
