% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QQ.R
\name{QQplot}
\alias{QQplot}
\title{Construct QQ Plot}
\usage{
QQplot(breaks, counts, sig = 0.05, color = "gray")
}
\arguments{
\item{breaks}{Breakpoints.}

\item{counts}{Number of test statistics exceeding the breakpoint.}

\item{sig}{Significance level for CIs.}

\item{color}{Point color.}
}
\description{
Construct QQ Plot
}
