% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadGeno.R
\name{ReadGeno}
\alias{ReadGeno}
\title{Read Genotypes from Plink}
\usage{
ReadGeno(m = 1, stem)
}
\arguments{
\item{m}{Number of loci to import.}

\item{stem}{Plink stem BED, BIM, FAM files.}
}
\description{
Reads a random block of $m$ snps from a plink data set.
}
