% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sampleFromData.R
\name{sampleFromData}
\alias{sampleFromData}
\title{Sample from a dataset}
\usage{
sampleFromData(data, size, replace = TRUE, ...)
}
\arguments{
\item{data}{The matrix or data.frame from which rows are to be
sampled.}

\item{size}{The number of samples to take.}

\item{replace}{Should values be replaced? By default takes the
value TRUE.}

\item{...}{Any other parameters to be passed to the sample
function}
}
\value{
Returns a dataset of the same type as the input data with \code{size} rows.
}
\description{
This function has been designed to sample from the rows of a two
dimensional data set returning all columns of the sampled rows.
}
\examples{
sampleFromData(airquality, 100)
}
\author{
Aimee Gott <agott@mango-solutions.com>
}

