% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{save.params}
\alias{save.params}
\title{Save parameter grid to output dir}
\usage{
save.params(paramgrid, outdir)
}
\arguments{
\item{outdir}{Output directory}

\item{paramlist}{Named list of parameters}

\item{exclusions}{Optional exclusions from the parameter grid}
}
\value{
Number of jobs total
}
\description{
Save parameter grid to output dir
}
