% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seeds.R
\name{set.stream.seed}
\alias{set.stream.seed}
\title{Set seed for a task}
\usage{
set.stream.seed(overall_seed, num_tasks, task_id)
}
\arguments{
\item{overall_seed}{Seed for entire process}

\item{num_tasks}{Number of job array tasks}

\item{task_id}{Task ID for this particular job}
}
\description{
Set seed for a task
}
