% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{setup.directories}
\alias{setup.directories}
\title{Setup output directories and write file description}
\usage{
setup.directories(basedir, desc)
}
\arguments{
\item{basedir}{Base directory to create the file in}

\item{desc}{Description to tag to the version of simulations}
}
\description{
Setup output directories and write file description
}
