% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simTools-package.R
\docType{package}
\name{simTools}
\alias{simTools}
\alias{simTools-package}
\title{simTools: A package for performing common simulation tasks}
\description{
This package provides a series of tools for common simulation
tasks such as sampling from a data frame and generating plots of
simulation experiments.
}
\author{
Aimee Gott \email{agott@mango-solutions.com}
}

