% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Skew.R
\name{Skew}
\alias{Skew}
\title{Estimate Skewness}
\usage{
Skew(y, s0 = 0, sig = 0.05, simple = F)
}
\arguments{
\item{y}{Observations.}

\item{s0}{Reference skewness.}

\item{sig}{Significance level.}

\item{simple}{If TRUE, returns estimated size and SE only.}
}
\value{
Matrix containing the estimated skewness, its standard error, the
  lower and upper confidence bounds, and a p value assessing the null
  hypothesis that the skewness is \code{s0}.
}
\description{
Estimates the skewness of a distribution using a vector of observations.
}
\examples{
\dontrun{
# Exponential sample
y = rexp(n=1e3);
# Test skewness = 2
Skew(y=y,s0=2);
}
}
