% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Smcmcclass.R
\name{Smcmc}
\alias{Smcmc}
\alias{as.Smcmc}
\alias{as.Smcmc.default}
\alias{is.mcmc}
\title{Smcmc class}
\usage{
Smcmc(data, batch.size = TRUE, stacked = TRUE, varnames = colnames(data))
}
\arguments{
\item{data}{: a list of MCMC output matrices each with `nsim` rows and `p` columns}

\item{batch.size}{: logical argument, if true, calculates the batch size appropriate for this Markov chain. Setting to TRUE saves time in future steps.}

\item{stacked}{: recommended to be `TRUE`. logical argument, if true, stores a carefully stacked version of the MCMC output for use later.}

\item{varnames}{: a character string equal to the number of columns in \code{data}}
}
\value{
an Smcmc class object
}
\description{
Smcmc class for simulated data using Markov chain Monte Carlo
}
\examples{
# Producing Markov chain
chain <- matrix(0, nrow = 1e3, ncol = 1)
chain[1,] <- 0
err <- rnorm(1e3)
for(i in 2:1e3)
{
  chain[i,] <- .3*chain[i-1,] + err[i]
}
smcmc.obj <- Smcmc(chain)
}
