% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submit.R
\name{submit.jobs}
\alias{submit.jobs}
\title{Submit jobs to SLURM cluster}
\usage{
submit.jobs(njobs, shellname, scriptname, outdir, ...)
}
\arguments{
\item{njobs}{Number of jobs in array}

\item{shellname}{Shell script name}

\item{scriptname}{Name of script to run}

\item{outdir}{Output directory}

\item{...}{Additional arguments to pass to scriptname}
}
\description{
Submit jobs to SLURM cluster
}
