% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Smcmcclass.R
\name{summary.Smcmc}
\alias{summary.Smcmc}
\alias{print.summary.Smcmc}
\title{Summary function for Smcmc objects}
\usage{
\method{summary}{Smcmc}(x)

\method{print}{summary.Smcmc}(x, ...)
}
\arguments{
\item{x}{: a `Smcmc' class object}
}
\value{
return statistics of the all the dimensions(& chains) in Smcmc object
}
\description{
To show different statistics of the Smcmc object
}
\examples{
# Producing Markov chain
chain <- matrix(0, ncol = 1, nrow = 1e3)
chain[1,] <- 0
err <- rnorm(1e3)
for(i in 2:1e3)
{
  chain[i,] <- .3*chain[i-1,] + err[i]
}
chain <- Smcmc(list(chain))
summary(chain)

}
