% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{traceplot}
\alias{traceplot}
\title{Trace Plot for Markov chain Monte Carlo}
\usage{
traceplot(x, fast = TRUE, which = NULL, xlim = NULL, ylim = NULL, main = NULL, xlab = "Iteration", 
                 ylab = NULL, alpha.f = 0.9, legend = TRUE, ask = dev.interactive(), col = c("palevioletred3","steelblue3","tan3","dimgrey","palegreen3"), ...)
}
\arguments{
\item{x}{: an `Smcmc' class object or a list of Markov chains or a Markov chain matrix or a vector.}

\item{fast}{: a Boolean argument that will be set to TRUE by default, to make plots faster.}

\item{which}{: if we want full size trace plots of specific dimensions of chain, we can pass a vector of respective dimension/components.}

\item{xlim}{: range of x-axis}

\item{ylim}{: range of y-axis}

\item{main}{: usual heading for plot}

\item{xlab}{: labels of x-axis}

\item{ylab}{: labels of y-axis,it should be a vector of length equal to dimension of chain.}

\item{alpha.f}{: To fix the opacity of lines as per user convenience, by default it is 0.9.}

\item{legend}{: Boolean argument, for making legend or not.}

\item{ask}{: activating interactive plots}

\item{col}{: color vector for multiple chains}

\item{...}{: Other arguments}
}
\value{
Returns the Trace Plots of Markov Chain(s)
}
\description{
traceplot is a graphical tool commonly used in Bayesian statistics and Markov Chain Monte Carlo(MCMC) methods to diagnose the convergence and mixing properties of a chain.
}
\examples{
# example code
# Defining a function to produce Markov chain with dimension p and size n
MakeChain <- function(p, n , h = .5)
{
  chain <- matrix(0, nrow = n,ncol = p)
  for (i in 2:n) 
  {
    prop <- chain[i-1, ] + rnorm(p, mean = 0, sd = h)
    log.ratio <- sum(dnorm(prop, log = TRUE) - dnorm(chain[i-1, ], log = TRUE))
    if(log(runif(1)) < log.ratio) 
    {chain[i, ] <- prop}
    else{chain[i, ] <- chain[i - 1, ]}
  }
  v = vector(length = p)
  for(i in 1:p){v[i] = paste("Comp ",i)}
  colnames(chain) = v
  return(chain)
}

chain1 <- MakeChain(p=4,n=1000)
chain2 <- MakeChain(p=4,n=1000)
chain3 <- MakeChain(p=4,n=1000)
out <- Smcmc(list(chain1,chain2,chain3))
traceplot(out)


chain1 <- MakeChain(p=6,n=1000)
chain2 <- MakeChain(p=6,n=1000)
chain3 <- MakeChain(p=6,n=1000)
out <- Smcmc(list(chain1,chain2,chain3))
traceplot(out)

}
