% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Smcmcclass.R
\name{convert2Smcmc}
\alias{convert2Smcmc}
\title{Covert to Smcmc Object}
\usage{
convert2Smcmc(x)
}
\arguments{
\item{x}{: a object belongs from any of "mcmc.list", "stanfit", "rstan", "array", "matrix" classes.}
}
\value{
return Smcmc object having same chain(s)
}
\description{
To covert different MCMC objects to Smcmc object
}
\examples{

library(StanHeaders)
library(rstan)

stanmodelcode <- "
data { 
 int<lower=0> n;
 vector[2] x[n];
 }
 parameters {
 vector[2] mu;
 vector<lower=0>[2] lambda;
 real<lower=-1,upper=1> r;
 } 
 transformed parameters {
 vector<lower=0>[2] sigma;
 cov_matrix[2] T;
 sigma[1] <- inv_sqrt(lambda[1]);
 sigma[2] <- inv_sqrt(lambda[2]);
 T[1,1] <- square(sigma[1]);
 T[1,2] <- r * sigma[1] * sigma[2];
 T[2,1] <- r * sigma[1] * sigma[2];
 T[2,2] <- square(sigma[2]);
 }
 model {
 mu ~ normal(0, inv_sqrt(.001));
 lambda ~ gamma(.001, .001);
 x ~ multi_normal(mu, T);
 }"

 x <- matrix(c(0.8,102,1.0,98,0.5,100,0.9,105,0.7,103,0.4,110, 1.2,99, 1.4,87,0.6,113,1.1,89,
                1.3,93), nrow=11, ncol=2, byrow=TRUE) 
                
 n <- nrow(x) # number of people/units measured

data <- list(x=x, n=n) # to be passed on to Stan
myinits <- list(list(r=0))

# parameters to be monitored: 
parameters <- c("r")

# The following command calls Stan with specific options.
# For a detailed description type "?rstan".
fit <- stan(model_code=stanmodelcode,data=data,init=myinits,iter=1000,chains=1,thin=1)

convert2Smcmc(fit)

}
