% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleFromData.R
\name{sampleFromData}
\alias{sampleFromData}
\title{Sample From Data}
\usage{
sampleFromData(data, size, replace = TRUE, ...)
}
\arguments{
\item{data}{A data frame}

\item{size}{integer. The number of rows to return in the sample}

\item{replace}{logical. If \code{TRUE} then sample with replacement}

\item{...}{passed on to the sample function}
}
\value{
A data frame subsampled from \code{data}
}
\description{
This function returns a sample from a data frame.
}
\details{
Some details about this function. Longer prose
that is too long for a description
}
\examples{
irisSmall <- sampleFromData(iris, size = 10, replace = FALSE)
nrow(irisSmall)

\dontrun{
print(nothing)
}
}
\author{
Doug Ashton <dashton@mango-solutions.com>
}

