% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simTools-package.R
\docType{package}
\name{simTools}
\alias{simTools}
\alias{simTools-package}
\title{simTools: A package for performing comming simulation tasks}
\description{
This package description will be filled in during the break
}
\author{
Doug Ashton
}

