% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Smcmcclass.R
\name{Smcmc}
\alias{Smcmc}
\alias{as.Smcmc}
\alias{as.Smcmc.default}
\alias{is.mcmc}
\title{Smcmc class}
\usage{
Smcmc(data, batch.size = FALSE, varnames = colnames(data))
}
\arguments{
\item{data}{: an MCMC output matrix with nsim rows and p columns}

\item{batch.size}{: logical vector, if true, calculates the batch size appropriate for this Markov chain. Setting to TRUE saves time in future steps.}

\item{varnames}{: a character string equal to the number of columns in \code{data}}
}
\value{
an Smcmc class object
}
\description{
Smcmc class for simulated data using Markov chain Monte Carlo
}
\examples{
# Producing Markov chain
chain <- numeric(length = 1e3)
chain[1] <- 0
err <- rnorm(1e3)
for(i in 2:1e3)
{
  chain[i] <- .3*chain[i-1] + err[i]
}
smcmc.obj <- Smcmc(chain)
}
