% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R

\name{traceplot}
\alias{traceplot}
\title{
Trace Plot for Markov chain Monte Carlo
}
\description{
traceplot is a graphical tool commonly used in Bayesian statistics and Markov Chain Monte Carlo(MCMC) methods to diagnose the convergence and mixing properties of a chain. 
}
\usage{
traceplot(x, fast = FALSE, xlim = NULL, ylim = NULL,component = NULL,
          xlab = "Iteration",auto.layout = TRUE,
          ask = dev.interactive(),
          col = c("palevioletred3","steelblue3","tan3","lightsteelblue3",
                  "springgreen2","skyblue3","khaki3",
                  "lightpink1","palegreen3","thistle3","aquamarine3",
                  "dimgrey","tomato3"))
}
\arguments{
  \item{x}{: an `Smcmc' class object or a list of Markov chains or a Markov chain matrix}
  
  \item{fast}{: a boolean argument that will be auto set to TRUE when chain size is larger than 1e6}
  
  \item{xlim}{: range of x-axis}
  
  \item{ylim}{: range of y-axis}
  
  \item{component}{: a vector of integers indicating which components' ACF plots are needed. By default all components are drawn.}
  
  \item{xlab}{: lables of x-axis}
  
  \item{auto.layout}{: logical argument for an automatic layout of plots}
  
  \item{ask}{: activating interactive plots}
  
  \item{col}{: color vector for multiple chains}
}

\value{
Returns the Trace Plots of Markov Chain(s).
}

\examples{

# Defining a function to produce Markov chain with dimension p and size n
MakeChain <- function(p, n , h = .5)
{
chain <- matrix(0, nrow = n,ncol = p)
for (i in 2:n) 
{
  prop <- chain[i-1, ] + rnorm(p, mean = 0, sd = h)
  log.ratio <- sum(dnorm(prop, log = TRUE) - dnorm(chain[i-1, ], log = TRUE))
  if(log(runif(1)) < log.ratio) 
    {chain[i, ] <- prop}
  else{chain[i, ] <- chain[i - 1, ]}
}
  v = vector(length = p)
  for(i in 1:p){v[i] = paste("Comp ",i)}
  colnames(chain) = v
  return(chain)
}


chain1 <- MakeChain(p=4,n=1000)
chain2 <- MakeChain(p=4,n=1000)
chain3 <- MakeChain(p=4,n=1000)
out <- Smcmc(list(chain1,chain2,chain3))
traceplot(out)


chain1 <- MakeChain(p=6,n=1000)
chain2 <- MakeChain(p=6,n=1000)
chain3 <- MakeChain(p=6,n=1000)
out <- Smcmc(list(chain1,chain2,chain3))
traceplot(out)

}

